Tweaks to make Xcode 6 beta happy

+++ Feather/Sources/Database Packages/MPDatabasePackageController.h
@@ -52,7 +52,7 @@ typedef enum MPDatabasePackageControllerErrorCode
 /** The database server for this database package. */
 @property (strong, readonly) CBLManager *server;
 
-@property (readonly) NSUInteger serverQueueToken;
+@property (readonly) unsigned long long serverQueueToken;
 
 /** The base remote URL for the document package. NOTE! An abstract method. */
 @property (strong, readonly) NSURL *remoteURL;
